requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
openai==1.30.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
aiofiles==23.2.1
